modules/moltbot/requirements.txt
# Moltbot News Crawler - Python Dependencies
# 安裝方式: pip install -r requirements.txt

# 爬蟲模塊
requests>=2.31.0
playwright>=1.40.0
beautifulsoup4>=4.12.0

# 關鍵字分析
jieba>=0.42.1

# API 連接器
openai>=1.10.0
google-generativeai>=0.3.0

# LINE Bot
line-bot-sdk>=3.5.0
flask>=3.0.0

# 測試
pytest>=7.4.0

# 環境變數
python-dotenv>=1.0.0

